requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyMuPDF==1.24.9
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
urllib3==2.0.7
beautifulsoup4==4.12.2
chardet==5.2.0
python-docx==0.8.11
gunicorn==21.2.0
pywin32==306; platform_system == "Windows"
cohere==5.5.8
waitress==2.1.2
gunicorn==21.2.0